fix: add new db
