--- alembic/versions/085c329de782_change_chat_id_to_string.py
@@ -1,30 +1,0 @@
-"""Change chat id to string
-
-Revision ID: 085c329de782
-Revises: 29aa49641cef
-Create Date: 2024-06-27 11:32:59.396202
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '085c329de782'
-down_revision: Union[str, None] = '29aa49641cef'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
