--- alembic/versions/151753512221_add_lead_name_column_to_chat.py
@@ -1,35 +1,0 @@
-"""Add lead_name column to Chat
-
-Revision ID: 151753512221
-Revises: e6ffb480e843
-Create Date: 2024-06-23 22:25:10.882433
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = "151753512221"
-down_revision: Union[str, None] = "e6ffb480e843"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Add lead_name column as nullable first
-    op.add_column("chats", sa.Column("lead_name", sa.String(length=100), nullable=True))
-    # Populate the new column with a default value
-    op.execute("UPDATE chats SET lead_name = 'default_name' WHERE lead_name IS NULL")
-    # Alter the column to be non-nullable
-    op.alter_column("chats", "lead_name", nullable=False)
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("chats", "lead_name")
-    # ### end Alembic commands ###
