--- alembic/versions/29aa49641cef_add_registration_date.py
@@ -1,30 +1,0 @@
-"""add registration_date
-
-Revision ID: 29aa49641cef
-Revises: ffca00d45f26
-Create Date: 2024-06-27 10:53:21.644770
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '29aa49641cef'
-down_revision: Union[str, None] = 'ffca00d45f26'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('registration_date', sa.DateTime(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'registration_date')
-    # ### end Alembic commands ###
