--- alembic/versions/58f28d679fe1_sync_with_database.py
@@ -1,30 +1,0 @@
-"""sync with database
-
-Revision ID: 58f28d679fe1
-Revises: 42dbff0214c2
-Create Date: 2024-06-09 13:40:37.671743
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '58f28d679fe1'
-down_revision: Union[str, None] = '42dbff0214c2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
