+++ alembic/versions/66f7775bea6d_initial.py
@@ -1,0 +1,73 @@
+"""initial
+
+Revision ID: 66f7775bea6d
+Revises: 
+Create Date: 2024-06-27 19:52:05.537667
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = '66f7775bea6d'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('name', sa.String(length=100), nullable=False),
+    sa.Column('has_profile', sa.Boolean(), nullable=True),
+    sa.Column('words', sa.BigInteger(), nullable=True),
+    sa.Column('registration_date', sa.DateTime(), nullable=True),
+    sa.Column('auth_status', sa.String(length=20), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('chats',
+    sa.Column('id', sa.String(length=255), nullable=False),
+    sa.Column('name', sa.String(length=100), nullable=False),
+    sa.Column('words', sa.BigInteger(), nullable=True),
+    sa.Column('status', postgresql.ENUM('pending', 'sold', 'declined', name='chatstatus'), nullable=False),
+    sa.Column('lead_id', sa.BigInteger(), nullable=True),
+    sa.Column('lead_name', sa.String(length=100), nullable=False),
+    sa.ForeignKeyConstraint(['lead_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('agreed_users',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('chat_id', sa.String(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('user_id', 'chat_id')
+    )
+    op.create_table('agreed_users_chats',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('chat_id', sa.String(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('user_id', 'chat_id')
+    )
+    op.create_table('users_chats',
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('chat_id', sa.String(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('user_id', 'chat_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('users_chats')
+    op.drop_table('agreed_users_chats')
+    op.drop_table('agreed_users')
+    op.drop_table('chats')
+    op.drop_table('users')
+    # ### end Alembic commands ###
