--- alembic/versions/74ed918008e7_replace_int_with_bigint.py
@@ -1,30 +1,0 @@
-"""Replace Int with BigInt
-
-Revision ID: 74ed918008e7
-Revises: bd67ccd21c3f
-Create Date: 2024-06-10 11:26:35.591583
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '74ed918008e7'
-down_revision: Union[str, None] = 'bd67ccd21c3f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
