--- alembic/versions/7e0e51bac25d_add_private_id_and_is_private.py
@@ -1,32 +1,0 @@
-"""add private_id and is_private
-
-Revision ID: 7e0e51bac25d
-Revises: 085c329de782
-Create Date: 2024-06-27 18:30:56.303903
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '7e0e51bac25d'
-down_revision: Union[str, None] = '085c329de782'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('chats', sa.Column('is_private', sa.Boolean(), nullable=True))
-    op.add_column('chats', sa.Column('private_id', sa.String(length=200), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('chats', 'private_id')
-    op.drop_column('chats', 'is_private')
-    # ### end Alembic commands ###
