--- alembic/versions/83f918ec405e_fix_typo_in_back_populates_for_agreed_.py
@@ -1,30 +1,0 @@
-"""Fix typo in back_populates for agreed_users
-
-Revision ID: 83f918ec405e
-Revises: 8813c6adc9f9
-Create Date: 2024-06-09 22:25:15.400346
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '83f918ec405e'
-down_revision: Union[str, None] = '8813c6adc9f9'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
