--- alembic/versions/8813c6adc9f9_add_full_text_column_to_chat.py
@@ -1,30 +1,0 @@
-"""Add full_text column to chat
-
-Revision ID: 8813c6adc9f9
-Revises: f065b87e79de
-Create Date: 2024-06-09 19:51:24.280403
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '8813c6adc9f9'
-down_revision: Union[str, None] = 'f065b87e79de'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('chats', sa.Column('full_text', sa.Text(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('chats', 'full_text')
-    # ### end Alembic commands ###
