--- alembic/versions/9547e28e8308_add_agreed_users_table.py
@@ -1,68 +1,0 @@
-"""Add agreed_users table
-
-Revision ID: 9547e28e8308
-Revises: 58f28d679fe1
-Create Date: 2024-06-09 17:58:37.407721
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '9547e28e8308'
-down_revision: Union[str, None] = '58f28d679fe1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('chats',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=100), nullable=False),
-    sa.Column('words', sa.Integer(), nullable=True),
-    sa.Column('status', postgresql.ENUM('pending', 'sold', 'declined', name='chatstatus'), nullable=False),
-    sa.Column('lead_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['lead_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('agreed_users',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('chat_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('user_id', 'chat_id')
-    )
-    op.create_table('users_chats',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('chat_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('user_id', 'chat_id')
-    )
-    op.add_column('users', sa.Column('has_profile', sa.Boolean(), nullable=True))
-    op.add_column('users', sa.Column('words', sa.Integer(), nullable=True))
-    op.alter_column('users', 'name',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=False)
-    op.drop_constraint('users_email_key', 'users', type_='unique')
-    op.drop_column('users', 'email')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('email', sa.VARCHAR(length=100), autoincrement=False, nullable=True))
-    op.create_unique_constraint('users_email_key', 'users', ['email'])
-    op.alter_column('users', 'name',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=True)
-    op.drop_column('users', 'words')
-    op.drop_column('users', 'has_profile')
-    op.drop_table('users_chats')
-    op.drop_table('agreed_users')
-    op.drop_table('chats')
-    # ### end Alembic commands ###
