--- alembic/versions/bb7edb716774_add_new_table_agreed_users_chats.py
@@ -1,36 +1,0 @@
-"""Add new table agreed_users_chats
-
-Revision ID: bb7edb716774
-Revises: 158a52fbe4a7
-Create Date: 2024-06-09 23:02:13.176432
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bb7edb716774'
-down_revision: Union[str, None] = '158a52fbe4a7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('agreed_users_chats',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('chat_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['chat_id'], ['chats.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('user_id', 'chat_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('agreed_users_chats')
-    # ### end Alembic commands ###
