--- alembic/versions/bd67ccd21c3f_fix_typo.py
@@ -1,30 +1,0 @@
-"""Fix typo
-
-Revision ID: bd67ccd21c3f
-Revises: bb7edb716774
-Create Date: 2024-06-09 23:17:12.318548
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bd67ccd21c3f'
-down_revision: Union[str, None] = 'bb7edb716774'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
