--- alembic/versions/dc696ca83a39_update_columns_to_bigint.py
@@ -1,30 +1,0 @@
-"""Update columns to BIGINT
-
-Revision ID: dc696ca83a39
-Revises: 74ed918008e7
-Create Date: 2024-06-10 11:53:41.836175
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'dc696ca83a39'
-down_revision: Union[str, None] = '74ed918008e7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
