--- alembic/versions/ed2b23ed3aaa_update_columns_to_bigint.py
@@ -1,30 +1,0 @@
-"""Update columns to BIGINT
-
-Revision ID: ed2b23ed3aaa
-Revises: dc696ca83a39
-Create Date: 2024-06-10 12:00:50.294734
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ed2b23ed3aaa'
-down_revision: Union[str, None] = 'dc696ca83a39'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
