--- alembic/versions/f065b87e79de_remove_chats_agreed.py
@@ -1,30 +1,0 @@
-"""Remove chats_agreed
-
-Revision ID: f065b87e79de
-Revises: 9547e28e8308
-Create Date: 2024-06-09 19:22:20.619469
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f065b87e79de'
-down_revision: Union[str, None] = '9547e28e8308'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
