--- alembic/versions/ffca00d45f26_add_auth_status_to_user_table.py
@@ -1,30 +1,0 @@
-"""Add auth_status to user table
-
-Revision ID: ffca00d45f26
-Revises: 151753512221
-Create Date: 2024-06-26 14:28:54.737952
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ffca00d45f26'
-down_revision: Union[str, None] = '151753512221'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('auth_status', sa.String(length=20), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'auth_status')
-    # ### end Alembic commands ###
